% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_get_admin2_boundaries.R
\name{f_get_admin2_boundaries}
\alias{f_get_admin2_boundaries}
\title{Get admin2 polygons from API}
\usage{
f_get_admin2_boundaries(ISO3, simplify = TRUE, dTolerance = 500)
}
\arguments{
\item{ISO3}{ISO3 code of country}

\item{simplify}{Logical: whether to simplify or not}

\item{dTolerance}{parameter passed to \code{\link[sf:geos_unary]{sf::st_simplify()}}}
}
\value{
sf object
}
\description{
Queries the gis.unhcr.org sever to return admin2 shape files for a specific
country. Some of the files returned are heavy, and the server also seems to
fail or time out every now and then. Also, some admin2 codes are not in the
expected format.
}
\examples{
## The test is actually the QA process! 
library(countrycode)
ctr <- countrycode::codelist |>
       dplyr::filter( ! is.na(iso3c) )  |>
       dplyr::filter(continent == "Americas" ) |>
       dplyr::select(country.name.en, iso3c)
ctr
ctr <- ctr |>
       dplyr::select(iso3c)  |>
       dplyr::pull()  


# 
# BRA <- f_get_admin2_boundaries(ISO3 = "BRA", simplify = TRUE, dTolerance = 500)
# ECU <- f_get_admin2_boundaries(ISO3 = "ECU", simplify = TRUE, dTolerance = 500)
# GTM <- f_get_admin2_boundaries(ISO3 = "GTM", simplify = TRUE, dTolerance = 500)
# CRI <- f_get_admin2_boundaries(ISO3 = "CRI", simplify = TRUE, dTolerance = 500)
# CHL <- f_get_admin2_boundaries(ISO3 = "CHL", simplify = TRUE, dTolerance = 500)
# COL <- f_get_admin2_boundaries(ISO3 = "COL", simplify = TRUE, dTolerance = 500)
# MEX <- f_get_admin2_boundaries(ISO3 = "MEX", simplify = TRUE, dTolerance = 500)


for ( code  in ctr )
 { t <- f_get_admin2_boundaries(ISO3 = code, simplify = TRUE, dTolerance = 500) }
}
