% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_get_admin2_cod.R
\name{f_get_admin2_cod}
\alias{f_get_admin2_cod}
\title{Get admin2 polygons from API}
\usage{
f_get_admin2_cod(ISO3, simplify = TRUE, dTolerance = 500)
}
\arguments{
\item{ISO3}{ISO3 code of country}

\item{simplify}{Logical: whether to simplify or not}

\item{dTolerance}{parameter passed to \code{\link[sf:geos_unary]{sf::st_simplify()}}}
}
\value{
sf object
}
\description{
Queries the COD API sever to return admin2 geometry and pcode for a specific
country based on iso name.
As the resulting data is expected to be used in a shinyApp, the geomtry is simplified
}
\examples{
#GTM <- f_get_admin2_cod("GTM", simplify = TRUE, dTolerance = 500)
}
